refactor: generalize GraphQL error handling for RPC errors
- Move GraphQLErrorDetails and ErrorWithGraphQL types to utils/errors.ts
- Add request() wrapper in utils/graphql.ts to handle RPC errors in GraphQL responses
- Simplify hooks/payments/crypto.ts by removing ~60 lines of repetitive error handling
- All GraphQL operations that may return RPC errors now use the centralized wrapper

This provides consistent error handling across the codebase and makes it easy
to extend to other GraphQL operations that interact with backend services.
